save changes

+++ app/verification/service.py
@@ -23,7 +23,6 @@ class VerificationService:
         return generate_otp() if type == 'otp' else generate_jwt_token({"user_id": user_id})
 
     def _validate_verification(self, verification: Verification, user_id: UUID, identifier: str) -> None:
-        if not verification: raise HTTPException(status_code=401, detail=ERROR_INVALID_TOKEN)
         if verification.user_id != user_id: raise HTTPException(status_code=401, detail=ERROR_TOKEN_DOES_NOT_BELONG_TO_USER)
         if verification.identifier != identifier: raise HTTPException(status_code=401, detail=ERROR_INVALID_IDENTIFIER)
         if verification.expires_at < datetime.now(timezone.utc): raise HTTPException(status_code=401, detail=ERROR_TOKEN_EXPIRED)
@@ -31,10 +30,11 @@ class VerificationService:
     async def create_verification(self, identifier: str, user_id: UUID, type: Literal['otp', 'link'] = 'otp') -> Verification:
         token = self._generate_token(type, user_id)
         verification = Verification(user_id=user_id, identifier=identifier, token=token)
-        return await self.verification_repository.create(verification, commit=False)
+        return await self.verification_repository.create(verification)
 
     async def verify(self, token: str, identifier: str, user_id: UUID) -> bool:
         verification = await self.verification_repository.get_by_token(token)
+        if not verification: return False
         self._validate_verification(verification, user_id, identifier)
         await self.verification_repository.delete(verification.id)
         return True
